requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.8.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
openpyxl==3.1.5
pillow==11.2.1
pycparser==2.22
PyJWT==2.10.1
pypdf==5.5.0
python-docx==1.1.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
urllib3==2.4.0
xhtml2pdf==0.2.17
inflection==0.5.1
pytz==2025.2
uritemplate==4.1.1
packaging==25.0
lxml
typing-extensions